review refactors
